[SYCL][Graph] Addressing PR comments
- Rename some tests for better clarity
- Unify more subgraph tests
- Fix various small issues
