+++ sycl/test-e2e/Graph/Explicit/add_node_while_recording.cpp
