+++ sycl/test-e2e/Graph/Inputs/buffer_copy_host2target.cpp
@@ -35,4 +35,6 @@ int main() {
   for (size_t i = 0; i < Size; i++) {
     assert(ReferenceA[i] == HostAccA[i]);
   }
+
+  return 0;
 }
