+++ sycl/test-e2e/Graph/Inputs/buffer_copy_host2target_offset.cpp
@@ -39,4 +39,6 @@ int main() {
   for (size_t i = 0; i < Size + Offset; i++) {
     assert(ReferenceA[i] == HostAccA[i]);
   }
+
+  return 0;
 }
