+++ sycl/test-e2e/Graph/Inputs/buffer_copy_target2host.cpp
@@ -35,4 +35,6 @@ int main() {
     assert(ReferenceA[i] == DataA[i]);
     assert(ReferenceB[i] == DataB[i]);
   }
+
+  return 0;
 }
