+++ sycl/test-e2e/Graph/Inputs/buffer_copy_target2host_offset.cpp
@@ -39,4 +39,6 @@ int main() {
     assert(ReferenceA[i] == DataA[i]);
     assert(ReferenceB[i] == DataB[i]);
   }
+
+  return 0;
 }
