+++ sycl/test-e2e/Graph/Inputs/stream.cpp
@@ -37,20 +37,3 @@ int main() {
 
   return 0;
 }
-
-// CHECK-DAG: Val: 1
-// CHECK-DAG: Val: 2
-// CHECK-DAG: Val: 3
-// CHECK-DAG: Val: 4
-// CHECK-DAG: Val: 5
-// CHECK-DAG: Val: 6
-// CHECK-DAG: Val: 7
-// CHECK-DAG: Val: 8
-// CHECK-DAG: Val: 9
-// CHECK-DAG: Val: 10
-// CHECK-DAG: Val: 11
-// CHECK-DAG: Val: 12
-// CHECK-DAG: Val: 13
-// CHECK-DAG: Val: 14
-// CHECK-DAG: Val: 15
-// CHECK-DAG: Val: 16
