+++ sycl/test-e2e/Graph/RecordReplay/sub_graph_two_parent_graphs.cpp
@@ -6,6 +6,9 @@
 //
 // CHECK-NOT: LEAK
 
+// XFAIL: *
+// Subgraph doesn't work properly in second parent graph
+
 #define GRAPH_E2E_RECORD_REPLAY
 
-#include "../Inputs/add_nodes_after_finalize.cpp"
+#include "../Inputs/sub_graph_two_parent_graphs.cpp"
